dist non iid param added

+++ src/options.py
@@ -60,7 +60,7 @@ def args_parser():
     parser.add_argument('--unequal', type=int, default=0,
                         help='whether to use unequal data splits for  \
                         non-i.i.d setting (use 0 for equal splits)')
-    parser.add_argument('--dist_noniid', type=int, default=0,
+    parser.add_argument('--dist_noniid', type=float, default=0,
                         help='whether to use distribution-based label imbalande for  \
                         non-i.i.d setting (use 0 for equal splits)')
     parser.add_argument('--stopping_rounds', type=int, default=10,
